fix: split api and web tests, fix flaky tests
