+++ package.json
@@ -9,7 +9,8 @@
   "license": "ISC",
   "devDependencies": {
     "@playwright/test": "^1.40.1",
-    "@types/node": "^20.10.6"
+    "@types/node": "^20.10.6",
+    "dotenv": "^16.3.1"
   },
   "dependencies": {
     "@zebrunner/javascript-agent-playwright": "^0.1.0"
