+++ playwright.config.ts
@@ -1,8 +1,8 @@
-import { defineConfig, devices } from '@playwright/test';
+import { defineConfig, devices } from '@playwright/test'
+import 'dotenv/config'
 
 export default defineConfig({
   testDir: './tests',
-  testIgnore: 'checkTest.spec.ts',
   fullyParallel: true,
   forbidOnly: !!process.env.CI,
   retries: process.env.CI ? 2 : 1,
@@ -15,7 +15,7 @@ export default defineConfig({
         projectKey: 'YANA',
         server: {
           hostname: 'https://solvdinternal.zebrunner.com',
-          accessToken: '8bCivk6EZxgQ7a9XwUzBelqrJDV7f0opBe2c9H52eZQ3dnxgnn'
+          accessToken: process.env.ZEBRUNNER_ACCESS_TOKEN
         },
         launch: {
           displayName: "Playwright launch",
@@ -58,18 +58,26 @@ export default defineConfig({
   },
   projects: [
     {
+      name: 'api',
+      testMatch: 'api/*.spec.ts',
+    },
+
+    {
       name: 'chromium',
       use: { ...devices['Desktop Chrome'] },
+      testMatch: 'web/*.spec.ts',
     },
 
     {
       name: 'firefox',
       use: { ...devices['Desktop Firefox'] },
+      testMatch: 'web/*.spec.ts',
     },
 
     {
       name: 'webkit',
       use: { ...devices['Desktop Safari'] },
+      testMatch: 'web/*.spec.ts',
     },
   ],
 });
