+++ tests/web/testExploreSection.spec.ts
@@ -1,13 +1,13 @@
 import { test, expect } from '@playwright/test'
-import { HomePage } from '../pages/homePage'
-import { TrendingPage } from '../pages/trendingPage'
-import { MusicPage } from '../pages/musicPage'
-import { MoviesPage } from '../pages/moviesPage'
-import { LivePage } from '../pages/livePage'
-import { GamingPage } from '../pages/gamingPage'
-import { NewsPage } from '../pages/newsPage'
-import { SportsPage } from '../pages/sportsPage'
-import { PodcastsPage } from '../pages/podcastsPage'
+import { HomePage } from '../../pages/homePage'
+import { TrendingPage } from '../../pages/trendingPage'
+import { MusicPage } from '../../pages/musicPage'
+import { MoviesPage } from '../../pages/moviesPage'
+import { LivePage } from '../../pages/livePage'
+import { GamingPage } from '../../pages/gamingPage'
+import { NewsPage } from '../../pages/newsPage'
+import { SportsPage } from '../../pages/sportsPage'
+import { PodcastsPage } from '../../pages/podcastsPage'
 import { testRail, CurrentTest } from '@zebrunner/javascript-agent-playwright'
 
 test('check that the explore section titles on the home page and on each section page are the same', async ({ page }) => {
